feat(token-definitions): refactor token fetching logic and add support for token structure types
