+++ suite-common/token-definitions/src/tokenDefinitionsThunks.ts
@@ -1,19 +1,12 @@
-import { D, G } from '@mobily/ts-belt';
-import { decode, verify } from 'jws';
+import { D } from '@mobily/ts-belt';
 
 import { createThunk } from '@suite-common/redux-utils';
-import { NetworkSymbol, getCoingeckoId } from '@suite-common/wallet-config';
-import { getJWSPublicKey, isCodesignBuild } from '@trezor/env-utils';
+import { NetworkSymbol } from '@suite-common/wallet-config';
 import { TimerId } from '@trezor/type-utils';
 
-import {
-    JWS_SIGN_ALGORITHM,
-    TOKEN_DEFINITIONS_PREFIX_URL,
-    TOKEN_DEFINITIONS_SUFFIX_URL,
-} from './tokenDefinitionsConstants';
 import { selectNetworkTokenDefinitions } from './tokenDefinitionsSelectors';
-import { DefinitionType } from './tokenDefinitionsTypes';
-import { getSupportedDefinitionTypes } from './tokenDefinitionsUtils';
+import { DefinitionType, TokenStructureType } from './tokenDefinitionsTypes';
+import { fetchTokenDefinitions, getSupportedDefinitionTypes } from './tokenDefinitionsUtils';
 
 const TOKEN_DEFINITIONS_MODULE = '@common/wallet-core/token-definitions';
 
@@ -26,52 +19,12 @@ export const getTokenDefinitionThunk = createThunk(
         },
         { fulfillWithValue, rejectWithValue },
     ) => {
-        const { symbol, type } = params;
-        const coingeckoId = getCoingeckoId(symbol);
-
         try {
-            if (!coingeckoId) {
-                throw Error(
-                    'Cannot fetch token definitions for network without CoinGecko asset id!',
-                );
-            }
-
-            const env = isCodesignBuild() ? 'stable' : 'develop';
-
-            const response = await fetch(
-                `${TOKEN_DEFINITIONS_PREFIX_URL}/${env}/${coingeckoId}.simple.${type}.${TOKEN_DEFINITIONS_SUFFIX_URL}`,
+            const data = await fetchTokenDefinitions(
+                params.symbol,
+                params.type,
+                TokenStructureType.SIMPLE,
             );
-
-            if (!response.ok) {
-                throw Error(response.statusText);
-            }
-
-            const jws = await response.text();
-
-            const decodedJws = decode(jws);
-
-            if (!decodedJws) {
-                throw Error('Decoding of config failed');
-            }
-
-            const algorithmInHeader = decodedJws?.header.alg;
-            if (algorithmInHeader !== JWS_SIGN_ALGORITHM) {
-                throw Error(`Wrong algorithm in JWS config header: ${algorithmInHeader}`);
-            }
-
-            const authenticityPublicKey = getJWSPublicKey('token-definitions');
-
-            if (G.isNullable(authenticityPublicKey)) {
-                throw Error('Public key check token definitions authenticity was not found.');
-            }
-
-            const isAuthenticityValid = verify(jws, JWS_SIGN_ALGORITHM, authenticityPublicKey);
-
-            if (!isAuthenticityValid) {
-                throw Error('Config authenticity is invalid');
-            }
-
-            const data = JSON.parse(decodedJws.payload);
 
             return fulfillWithValue(data);
         } catch (error) {
